refactor: refine tflite
