--- nn_meter/builder/backends/openvino/utils/serial.py
@@ -1,2 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT license.
