+++ nn_meter/builder/backends/tflite/gpu.py
@@ -1,6 +1,8 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 import re
+from .tflite_runner import TFLiteRunner
+from .tflite_backend import TFLiteBackend
 from nn_meter.builder.utils.latency import Latency
 
 
@@ -163,3 +165,13 @@ class TFLiteGPUParser:
                 errors.append(match[1])
 
         return errors
+
+
+class TFLiteGPURunner(TFLiteRunner):
+    use_gpu = True
+
+
+class Backend(TFLiteBackend):
+    parser_class = TFLiteGPUParser
+    runner_class = TFLiteGPURunner
+
